for LDAP configuration, in documentation on https://doc.nuxeo.com/nxdoc/using-a-ldap-directory/#default-users-and-groups-configuration It is said that we can specify users as administrator throw : * Element defaultAdministratorId in XML configuration * nuxeo.ldap.defaultAdministratorId in the nuxeo.conf file
The search
https://github.com/nuxeo/nuxeo/search?l=XML&q=defaultAdministratorId
gives some examples of XML configuration, usually in unit test.
And we can see that node defaultAdministratorId can be repeated in the
XML configuration for extension
"org.nuxeo.ecm.platform.usermanager.UserService" on point "userManager".

But when using the nuxeo.conf, it seems impossible to specify multiple
user id as default administrator.

Also, according documentation
https://doc.nuxeo.com/nxdoc/configuration-parameters-index-nuxeoconf/#configuration-parameters-index
it explains that | is used to separate possible values. This is for the
documentation, but it was reused in this commit.

The change concerns the freemarker template :
* Takes value from nuxeo.ldap.defaultAdministratorId in the nuxeo.conf,
as it was done before
* Split the value according the |
* For each item, build a node defaultAdministratorId in the XML
configuration.